Make documentation of ObjectFactory::createObject more convenient
Put documentation on createObject and refer to it from getObjectFromSpec
instead of the other way around. With ObjectFactory available as a
service, the former is going to be much more common.

Also, duplicate some of the class documentation onto createObject.
That is the doc block most people will interact with, forcing them
to look up the class doc instead is inconvenient.

Change-Id: I7828fde53ff74b33bd85eff1f53fc1107b966513