Add support for optional services
Use the `optional_services` key in the object spec
to declare the optional services. These are treated
the same as the `services`, except that if the
ServiceContainer lacks the service called for,
the parameter is set to null instead of causes errors.

The PSR-11 Container interface specification requires
the `has` method, so there are no new dependencies
introduced.

Bug: T242050
Change-Id: I99408df6ed32a097b47f4192637444029990628a